WIP Engine + DSP + Synth

+++ kiro-audio/src/config.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, Clone)]
 pub struct AudioConfig {
-  pub sample_rate: usize,
+  pub sample_rate: u32,
   pub buffer_size: usize,
 }
 
 impl AudioConfig {
-  pub const DEFAULT_SAMPLE_RATE: usize = 44_100;
+  pub const DEFAULT_SAMPLE_RATE: u32 = 44_100;
   pub const DEFAULT_BUFFER_SIZE: usize = 256;
 }
